Rename Analysis.Spec.Lib to Analysis.Lib
