+++ plutus-core/plutus-core.cabal
@@ -416,8 +416,8 @@ test-suite untyped-plutus-core-test
   hs-source-dirs: untyped-plutus-core/test
   ghc-options:    -O2 -threaded -rtsopts -with-rtsopts=-N
   other-modules:
+    Analysis.Lib
     Analysis.Spec
-    Analysis.Spec.Lib
     DeBruijn.FlatNatWord
     DeBruijn.Scope
     DeBruijn.Spec
